Moving Browsersync tasks into a separated file
